@@ -3,22 +3,22 @@ import 'package:swagger_dart_code_generator/src/swagger_models/requests/paramete
 import 'package:swagger_dart_code_generator/src/swagger_models/requests/swagger_parameter_schema.dart';
 import 'package:swagger_dart_code_generator/src/swagger_models/requests/swagger_request_items.dart';
 
-part 'swagger_request_parameter.g2.dart';
+part 'swagger_request_parameter.g.dart';
 
 @JsonSerializable()
 class SwaggerRequestParameter {
   SwaggerRequestParameter(
-      {this.inParameter,
-      this.name,
-      this.description,
-      this.isRequired,
-      this.type,
+      {this.inParameter ='',
+      this.name = '',
+      this.description = '',
+      this.isRequired = false,
+      this.type = '',
       this.item,
-      this.collectionFormat,
+      this.collectionFormat = '',
       this.items,
       this.schema,
-      this.ref,
-      this.key});
+      this.ref = '',
+      this.key = ''});
 
   factory SwaggerRequestParameter.fromJson(Map<String, dynamic> json) =>
       _$SwaggerRequestParameterFromJson(json);
@@ -42,16 +42,16 @@ class SwaggerRequestParameter {
   String type;
 
   @JsonKey(name: 'item')
-  ParameterItem item;
+  ParameterItem? item;
 
   @JsonKey(name: 'collectionFormat')
   String collectionFormat;
 
   @JsonKey(name: 'schema')
-  SwaggerParameterSchema schema;
+  SwaggerParameterSchema? schema;
 
   @JsonKey(name: 'items')
-  SwaggerRequestItems items;
+  SwaggerRequestItems? items;
 
   @JsonKey(ignore: true)
   String key;
