@@ -1,10 +1,10 @@
 import 'package:json_annotation/json_annotation.dart';
 
-part 'parameter_item.g2.dart';
+part 'parameter_item.g.dart';
 
 @JsonSerializable()
 class ParameterItem {
-  ParameterItem({this.type, this.enumValues, this.defaultValue});
+  ParameterItem({this.type = '', this.enumValues = const[], this.defaultValue = ''});
 
   factory ParameterItem.fromJson(Map<String, dynamic> json) =>
       _$ParameterItemFromJson(json);
