@@ -1,13 +1,13 @@
 import 'package:swagger_dart_code_generator/src/swagger_models/requests/swagger_request_parameter.dart';
 
 class SwaggerComponents {
-  SwaggerComponents({this.parameters});
+  SwaggerComponents({this.parameters = const []});
 
   List<SwaggerRequestParameter> parameters;
 
   SwaggerComponents.fromJson(Map<String, dynamic> json)
       : parameters = json['parameters'] == null
-            ? null
+            ? []
             : mapParameters(json['parameters'] as Map<String, dynamic>);
 
   static List<SwaggerRequestParameter> mapParameters(
