@@ -1,11 +1,11 @@
 import 'package:json_annotation/json_annotation.dart';
 
-part 'swagger_info.g2.dart';
+part 'swagger_info.g.dart';
 
 @JsonSerializable()
 class SwaggerInfo {
   SwaggerInfo(
-      {this.description, this.version, this.title, this.termsOfService});
+      {this.description = '', this.version = '', this.title = '', this.termsOfService = ''});
 
   factory SwaggerInfo.fromJson(Map<String, dynamic> json) =>
       _$SwaggerInfoFromJson(json);
