@@ -3,7 +3,7 @@ import 'package:swagger_dart_code_generator/src/swagger_models/requests/swagger_
 import 'requests/swagger_request.dart';
 
 class SwaggerPath {
-  SwaggerPath({this.path, this.requests, this.parameters});
+  SwaggerPath({this.path = '', this.requests = const [], this.parameters = const []});
 
   String path;
   List<SwaggerRequest> requests;
