@@ -1,11 +1,11 @@
 import 'dart:core';
 import 'package:json_annotation/json_annotation.dart';
 
-part 'swagger_tag.g2.dart';
+part 'swagger_tag.g.dart';
 
 @JsonSerializable()
 class SwaggerTag {
-  SwaggerTag({this.name, this.description});
+  SwaggerTag({this.name = '', this.description = ''});
 
   factory SwaggerTag.fromJson(Map<String, dynamic> json) =>
       _$SwaggerTagFromJson(json);
